+++ lib/src/config_provider/config_types.dart
@@ -333,7 +333,7 @@ class CompilerOptsAuto {
   /// Extracts compiler options based on OS and config.
   List<String> extractCompilerOpts() {
     if (Platform.isMacOS && macIncludeStdLib) {
-      return getHeadersForMac();
+      return getCStandardLibraryHeadersForMac();
     }
 
     return [];